+++ src/Accelerators/NNPA/Dialect/ZHigh/ZHighOps/OpHelper.hpp
@@ -56,7 +56,7 @@ mlir::Value getMinusBcastConst(mlir::OpBuilder &builder, mlir::Location loc,
 
 // Get a constant tensor of given value and type.
 mlir::Value getConstantOfType(
-    mlir::OpBuilder &builder, mlir::Location loc, mlir::Type type, int64_t val);
+    mlir::OpBuilder &builder, mlir::Location loc, mlir::Type type, float val);
 
 /// True if at least one of the types is NHWC layout.
 bool oneIsOfNHWCLayout(mlir::Type t1, mlir::Type t2);
